backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6  # For file uploads

# Database (Python 3.13 stable versions)
sqlalchemy[asyncio]>=2.0.25
psycopg[binary]>=3.1.10  # Modern PostgreSQL driver
asyncpg>=0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.9  # Legacy driver for migrations if needed
alembic==1.13.1  # Migrations (future use)

# OpenAI
openai>=1.59.5

# Environment
python-dotenv>=1.0.0
pydantic-settings>=2.0.0  # For AuthConfig

# Authentication
python-jose[cryptography]>=3.3.0  # JWT token creation and validation
passlib>=1.7.4  # Password hashing
bcrypt>=4.0.0,<5.0.0  # Bcrypt backend for passlib (4.x for compatibility)
email-validator>=2.0.0  # Required for Pydantic EmailStr validation

# Security (Feature 8 HIPAA Compliance)
freezegun>=1.5.1  # Time testing for MFA and audit logs
pillow==10.2.0  # Image processing (qrcode dependency, already in exports)
pyotp>=2.9.0  # TOTP/MFA support
qrcode>=7.4.2  # QR code generation for MFA setup
user-agents>=2.2.0  # User agent parsing for device fingerprinting

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing async FastAPI
aiosqlite>=0.19.0  # Async SQLite driver for async test fixtures
faker>=20.0.0  # Realistic test data generation

# Audio pipeline dependencies (reuse from transcription pipeline)
pydub==0.25.1
audioop-lts>=0.2.1

# Rate limiting
slowapi==0.1.9

# Testing coverage
pytest-cov==4.1.0  # Code coverage reporting

# Background job scheduler
apscheduler==3.10.4  # Scheduled analytics aggregation jobs

# Export & Reporting dependencies (Feature 7)
# WeasyPrint system dependencies (must be installed separately):
# macOS: brew install cairo pango gdk-pixbuf libffi
# Ubuntu: apt-get install libpango-1.0-0 libpangoft2-1.0-0
weasyprint==60.1         # HTML to PDF conversion
python-docx==1.1.0       # DOCX generation
jinja2==3.1.2            # Template rendering
# pillow==10.2.0 moved to Security section (shared by qrcode and weasyprint)

# Performance testing (optional - install separately for load testing)
# pip install locust pytest-xdist psutil
# locust>=2.20.0  # Web-based load testing
# pytest-xdist>=3.5.0  # Parallel pytest execution
# psutil>=5.9.0  # System resource monitoring
